# backend/school-management/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management API",
        "description": "Documentación de la API para el sistema de gestión escolar",
        "contact": {
            "name": "Equipo de desarrollo CBTa71",
            "email": "soporte@cbta71.edu.mx"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Servidor local"
        },
        {
            "url": "https://nginx-production-728f.up.railway.app/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/api/v1/cards": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Listar métodos de pago del usuario autenticado",
                "description": "Obtiene la lista de tarjetas o métodos de pago asociados al usuario autenticado. Permite forzar actualización del caché.",
                "operationId": "getUserCards",
                "parameters": [
                    {
                        "name": "forceRefresh",
                        "in": "query",
                        "description": "Si es true, fuerza la actualización del caché.",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de métodos de pago obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cards": {
                                                    "description": "Lista de métodos de pago del usuario",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DisplayPaymentMethodResponse"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron métodos de pago.",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o ausente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cards"
                ],
                "summary": "Registrar un nuevo método de pago",
                "description": "Crea una sesión de Stripe Checkout para registrar una nueva tarjeta del usuario autenticado.",
                "operationId": "addUserCard",
                "responses": {
                    "201": {
                        "description": "Sesión de registro de método de pago creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cards": {
                                                    "description": "Url del checkout para agregar tarjeta",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/SetupCardResponse"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/cards/{paymentMethodId}": {
            "delete": {
                "tags": [
                    "Cards"
                ],
                "summary": "Eliminar un método de pago",
                "description": "Elimina un método de pago específico asociado al usuario autenticado.",
                "operationId": "deleteUserCard",
                "parameters": [
                    {
                        "name": "paymentMethodId",
                        "in": "path",
                        "description": "ID del método de pago a eliminar (por ejemplo, 'pm_1P7E89AjcPzVqRkV')",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pm_1P7E89AjcPzVqRkV"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Método de pago eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Método de pago eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Método de pago no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/users/{id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar los datos generales de un usuario",
                "operationId": "3cc4803b4b6d7fc974e8fb21b3b6b46b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DomainUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/DomainUser"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "El usuario ha sido actualizado con éxito."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/password": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualizar la contraseña de un usuario",
                "operationId": "2c04cec480a17c40c24162582ac82c29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password actualizada con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "DisplayPaymentMethodResponse": {
                "properties": {
                    "id": {
                        "description": "ID del método de pago",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "brand": {
                        "description": "Marca de la tarjeta",
                        "type": "string",
                        "example": "Visa",
                        "nullable": true
                    },
                    "masked_card": {
                        "description": "Número de tarjeta enmascarado",
                        "type": "string",
                        "example": "**** **** **** 4242",
                        "nullable": true
                    },
                    "expiration_date": {
                        "description": "Fecha de expiración de la tarjeta",
                        "type": "string",
                        "example": "12/25",
                        "nullable": true
                    },
                    "status": {
                        "description": "Estado del método de pago",
                        "type": "string",
                        "example": "active",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SetupCardResponse": {
                "properties": {
                    "id": {
                        "description": "ID de la configuración de la tarjeta",
                        "type": "string",
                        "example": "pm_01HXXXXXX",
                        "nullable": true
                    },
                    "url": {
                        "description": "URL para completar el setup de la tarjeta",
                        "type": "string",
                        "example": "https://example.com/setup/pm_01HXXXXXX",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DomainStudentDetail": {
                "description": "Detalles del estudiante",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 123
                    },
                    "career_id": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "n_control": {
                        "type": "integer",
                        "example": 20201234,
                        "nullable": true
                    },
                    "semestre": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "group": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "workshop": {
                        "type": "string",
                        "example": "Taller de programación",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "DomainUser": {
                "description": "Representa un usuario del sistema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "hashed_password"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+5215512345678"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "example": "1995-06-15",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "male",
                        "nullable": true
                    },
                    "curp": {
                        "type": "string",
                        "example": "PEMJ950615HDFRZN09"
                    },
                    "address": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Calle Falsa 123",
                            "Colonia Centro"
                        ],
                        "nullable": true
                    },
                    "stripe_customer_id": {
                        "type": "string",
                        "example": "cus_ABC123XYZ",
                        "nullable": true
                    },
                    "blood_type": {
                        "type": "string",
                        "example": "O+",
                        "nullable": true
                    },
                    "registration_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T12:34:56Z",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "activo",
                        "nullable": true
                    },
                    "studentDetail": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DomainStudentDetail"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}